+++ CHANGELOG.md
@@ -10,6 +10,8 @@ system. Old cached data will be converted to the new format automatically.
 
 ### Breaking changes
 
+* `getAll` and `getListing` now return an empty object instead of undefined when no items
+  match the query
 * `disconnect` event removed in favor of `disconnected`
 * `remoteStorage.claimAccess` removed in favor of `remoteStorage.access.claim`
 * `BaseClient#use` and `BaseClient#release` removed in favor of `BaseClient#cache`